This should be different PRs
- Restructure project, make API a proper AAR and pull shared things into barber-annotations
- Start resolved attr work
- Ditch bintray-release and move entirely to standard mavencentral structure (with snapshots!)
- Start 1.4.0-SNAPSHOT
