+++ README.md.template
@@ -172,6 +172,8 @@ dependencies {
 }
 ```
 
+Snapshots of the development version are available in [Sonatype's snapshots repository][snapshots].
+
 Proguard
 --------
 
@@ -193,3 +195,5 @@ License
     WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
     See the License for the specific language governing permissions and
     limitations under the License.
+
+ [snapshots]: https://oss.sonatype.org/content/repositories/snapshots/
