--- api/build.gradle
@@ -1,46 +1,0 @@
-apply plugin: 'com.android.library'
-apply plugin: 'bintray-release'
-
-def ci = "true".equals(System.getenv("CI"))
-
-android {
-    compileSdkVersion COMPILE_SDK
-    buildToolsVersion BUILD_TOOLS
-
-    defaultConfig {
-        versionCode VERSION_CODE
-        versionName VERSION_NAME
-        minSdkVersion 14
-        targetSdkVersion 23
-        consumerProguardFiles 'consumer-proguard-rules.pro'
-    }
-    compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_7
-        targetCompatibility JavaVersion.VERSION_1_7
-    }
-    lintOptions {
-        warning 'MissingTranslation'
-        abortOnError true
-        check 'NewApi', 'InlinedApi'
-        fatal 'NewApi', 'InlinedApi'
-        checkReleaseBuilds true
-
-        textReport ci
-        textOutput 'stdout'
-
-        htmlReport !ci
-        xmlReport !ci
-    }
-}
-
-publish {
-    userOrg = 'pandanomic'
-    groupId = GROUP_NAME
-    artifactId = 'barber-api'
-    version = VERSION_NAME
-    description = 'An annotation processor library for styling custom views.'
-    website = SITE_URL
-    bintrayUser = BINTRAY_USER
-    bintrayKey = BINTRAY_KEY
-    licences = ['Apache-2.0']
-}
