--- api/consumer-proguard-rules.pro
@@ -1,8 +1,0 @@
--keep class **$$Barbershop { *; }
--keep class io.sweers.barber.** { *; }
--keepclasseswithmembers class * {
-    @io.sweers.barber.* <fields>;
-}
--keepclasseswithmembers class * {
-    @io.sweers.barber.* <methods>;
-}
