+++ barber-annotations/gradle.properties
@@ -1,0 +1,3 @@
+POM_NAME=Barber Annotations
+POM_ARTIFACT_ID=barber-annotations
+POM_PACKAGING=jar
