+++ barber-api/gradle.properties
@@ -1,0 +1,3 @@
+POM_NAME=Barber API
+POM_ARTIFACT_ID=barber-api
+POM_PACKAGING=aar
