+++ barber-api/src/main/java/io/sweers/barber/package-info.java
@@ -1,0 +1,9 @@
+/**
+ * View attribute injection library for Android which generates the obtainStyledAttributes() and
+ * TypedArray boilerplate code for you at compile time.
+ * <p>
+ * No more handing to deal with context.obtainStyledAttributes(...) or manually retrieving values
+ * from the resulting {@link android.content.res.TypedArray TypedArray} instance. Just annotate your
+ * field or method with one of the annotations in the api package.
+ */
+package io.sweers.barber;
