+++ barber-compiler/gradle.properties
@@ -1,0 +1,3 @@
+POM_NAME=Barber Compiler
+POM_ARTIFACT_ID=barber-compiler
+POM_PACKAGING=jar
