+++ barber-compiler/src/main/resources/META-INF/services/javax.annotation.processing.Processor
@@ -1,0 +1,1 @@
+io.sweers.barber.BarberProcessor
