--- compiler/build.gradle
@@ -1,30 +1,0 @@
-apply plugin: 'java'
-apply plugin: 'bintray-release'
-
-sourceCompatibility = 1.7
-
-sourceSets {
-    main {
-        java {
-            srcDirs = ['src/main/java', '../api/src/main/java']
-        }
-    }
-}
-
-dependencies {
-    compile fileTree(dir: 'libs', include: ['*.jar'])
-    compile 'com.squareup:javapoet:1.4.0'
-    compile 'com.google.auto.service:auto-service:1.0-rc2'
-}
-
-publish {
-    userOrg = 'pandanomic'
-    groupId = GROUP_NAME
-    artifactId = 'barber-compiler'
-    version = VERSION_NAME
-    description = 'An annotation processor library for styling custom views.'
-    website = SITE_URL
-    bintrayUser = BINTRAY_USER
-    bintrayKey = BINTRAY_KEY
-    licences = ['Apache-2.0']
-}
