+++ gradle.properties
@@ -16,3 +16,16 @@
 # This option should only be used with decoupled projects. More details, visit
 # http://www.gradle.org/docs/current/userguide/multi_project_builds.html#sec:decoupled_projects
 # org.gradle.parallel=true
+GROUP=io.sweers.barber
+VERSION_NAME=1.4.0-SNAPSHOT
+
+POM_DESCRIPTION=An annotation processor library for styling custom views.
+POM_URL=https://github.com/hzsweers/Barber/
+POM_SCM_URL=https://github.com/hzsweers/Barber/
+POM_SCM_CONNECTION=scm:git:git://github.com/hzsweers/Barber.git
+POM_SCM_DEV_CONNECTION=scm:git:ssh://git@github.com/hzsweers/Barber.git
+POM_LICENCE_NAME=The Apache Software License, Version 2.0
+POM_LICENCE_URL=http://www.apache.org/licenses/LICENSE-2.0.txt
+POM_LICENCE_DIST=repo
+POM_DEVELOPER_ID=hzsweers
+POM_DEVELOPER_NAME=Zac Sweers
