+++ sample/build.gradle
@@ -1,48 +1,61 @@
+buildscript {
+    dependencies {
+        classpath rootProject.ext.androidPlugin
+        classpath rootProject.ext.aptPlugin
+    }
+}
+
 apply plugin: 'com.android.application'
 apply plugin: 'android-apt'
 
 android {
-    compileSdkVersion COMPILE_SDK
-    buildToolsVersion BUILD_TOOLS
+    compileSdkVersion rootProject.ext.compileSdkVersion
+    buildToolsVersion rootProject.ext.buildToolsVersion
 
     defaultConfig {
-        versionCode VERSION_CODE
-        versionName VERSION_NAME
-        applicationId "${GROUP_NAME}.sample"
-        minSdkVersion 14
-        targetSdkVersion 23
+        applicationId 'io.sweers.barber.sample'
+        versionCode 1
+        versionName "1.0.0"
+        minSdkVersion rootProject.ext.minSdkVersion
+        targetSdkVersion rootProject.ext.targetSdkVersion
     }
     buildTypes {
-        release {
-            minifyEnabled false
+        debug {
+            minifyEnabled true  // Make sure consumer proguard rules are working
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
     compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_7
-        targetCompatibility JavaVersion.VERSION_1_7
+        sourceCompatibility rootProject.ext.sourceCompatibilityVersion
+        targetCompatibility rootProject.ext.targetCompatibilityVersion
     }
     lintOptions {
-        textReport true
+        abortOnError true
+        check 'NewApi', 'InlinedApi'
+        fatal 'NewApi', 'InlinedApi'
+        checkReleaseBuilds true
+
+        textReport rootProject.ext.ci
         textOutput 'stdout'
-        abortOnError false
+
+        htmlReport !rootProject.ext.ci
+        xmlReport !rootProject.ext.ci
     }
     dexOptions {
-        preDexLibraries = false
+        preDexLibraries = !rootProject.ext.ci
+    }
+    packagingOptions {
+        exclude 'META-INF/services/javax.annotation.processing.Processor'
     }
 }
 
 dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
-    compile 'com.android.support:appcompat-v7:23.1.1'
+    compile rootProject.ext.supportAppCompat
 
-    apt project(':compiler')
-    compile project(':api')
+    apt project(':barber-compiler')
+    compile project(':barber-api')
 
-    testCompile 'junit:junit:4.12'
-    testCompile 'org.apache.maven:maven-ant-tasks:2.1.3' // Required for gradle 2.3 on macs for some reason
-    testCompile ('org.robolectric:robolectric:3.0') {
-        exclude group: 'commons-logging', module: 'commons-logging'
-        exclude group: 'org.apache.httpcomponents', module: 'httpclient'
-    }
+    testCompile rootProject.ext.junit
+    testCompile rootProject.ext.robolectric
 }
