+++ settings.gradle
@@ -1,1 +1,1 @@
-include ':sample', ':compiler', ':api'
+include ':sample', ':barber-compiler', ':barber-api', ':barber-annotations'
