--- version.properties
@@ -1,4 +1,0 @@
-code=6
-major=1
-minor=3
-patch=1
